Add RESX implementation too
